move the shunting yard code to it's own module
